requirements_streamlit.txt
streamlit
plotly
pandas
opencv-python
scikit-image
pillow
torch
torchvision
decord
matplotlib
scipy
numpy
tqdm
timm
einops
clip
sk-video
pyyaml
